Extend size of UtlString.appendNumber parameter so that it is big enough for 64bit numbers and add testcases to prove it
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14995 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXportLib/include/utl/UtlString.h
@@ -390,7 +390,7 @@ public:
     UtlString& append(const char c);
 
     /// Append a string representation of the value to the string
-    UtlString& appendNumber(int value, const char* format = "%d");
+    UtlString& appendNumber(Int64 value, const char* format = "%d");
     /**< MAX_NUMBER_STRING_SIZE is the maximum size text representation of the value allowed
      *   If this size is exceeded an error string is appended in place of the number
      */
